Change rationale and description of MD007
Originally MD007 (4 space list indent) was added purely because multi-markdown
required 4 space indents and wouldn't treat a 2-space indent as a nested list.
However, http://www.cirosantilli.com/markdown-styleguide/#indented-lists (and
the preceding section on that page) outlines another reason for doing this, so
the rule and explanation has been changed to address this.

Fixes #23