Add support for rule parameters
Refs #47

Rule parameters are settings that can be passed to rules in style files in
order to customize their behavior. The settings are provided in style files to
enforce a specific style.

This commit also changes how ruleset files are evaluated to be more sensible
(the block inside the rule is now instance_eval's as part of the Rule object
rather than as part of the ruleset object and building up parts of the rule as
needed).