Use single quotes to prevent early escaping
Before this commit, the use of double quotes and an escaped paren
and slash were being passed to the regex parsing engine as unescaped
characters, due to the double quotes handling the escaping.

Using single quotes, the escape is not handled in the string,
instead the regex engine sees the escaped characters and treats
them as literals, rather than part of the regex.