MD034: Bare URL used
Fixes #36

Note: this also introduces a new doc helper: find_type_element_except, which
will let you search for elements of a given type, but exclude descending into
children of specific element types. This was added because the
matching_text_element_lines method actually ended up searching the content of
links, which it explicitly said it didn't. You can now exclude children of
specific elements when searching for the content of text elements.