Probe for .mdlrc
Look up directories for `.mdlrc` before falling back to home. This
allows per project `.mdlrc`. The `-c` flag can be passed an alternate
filename to look for, unless it is a (relative or absolute) path, in
which case no probing will be performed and the path will be used as is.