Load .mdlrc from current directory by default
This change loads .mdlrc from the current directory by default, and will look
in parent directories until it finds a .mdlrc. If it doesn't find one, it will
try ~/.mdlrc.

If the user specifies a different config file with -c, then there won't be a
fallback to ~/.mdlrc if the specified config file isn't found.

Refs #124