Update changelog
The format of the changelog as well as the release process has been changed to
be closer to the example given on keepachangelog.com, as well as making it
easier when it comes time to make a new release. Merged pull requests won't be
called out in their own section, but will be referenced next to the thing they
fix/modify.