Fix lint MD039 checking for nodes inside link text
This fixes a bug when complicating the text inside a markdown link node.
For instance if you were to provide a smaller part of the text like
this:

  [blah *blargh* blah](...)

The kramdown parser would return three nodes instead of one as mdlint
would expect, and it would only take the first node 'blah ', detecting
the whitespace in the end, and raising an issue.

Closes #102