Add issue templates to help standardize requests
Adding issue templates to standardize the information we'd like contributors to include when they open issues. I've added four different types of templates to help gather the appropriate data for the various types of issues. I've created these after reviewing the issue history for this repo as well as the issue templates configured across other active projects.

These templates will be listed as options when a user opens a new issue via the github ui, as outlined [in the github documentation](https://help.github.com/en/articles/creating-issue-templates-for-your-repository). We can create a generic template for the organization include it as another option below these primary ones for the repository, see the [chef repo issues](https://github.com/chef/chef/issues/new/choose) for an example of this functionality.