Add a pull request template to help contributors
Adding a pull request template to standardize the information we'd like contributors to include when they open a pull request and updated the guidelines in `CONTRIBUTING.md` to reflect the changes in policies. I've added a single template to help gather the appropriate data and remind users of our requirements for code changes. I've created these after reviewing the pull request history for this repo as well as the pull request templates configured across other active projects.

This template will be listed as an option when a user opens a new pull request via the github ui, as outlined in the github documentation. We can create a directory with a different template for the different categories of pull requests, similar to the issue templates if we determine it's necessary.