Add a pull request template to help contributors (#276)
* Add a pull request template to help contributors

Adding a pull request template to standardize the information we'd like contributors to include when they open a pull request and updated the guidelines in `CONTRIBUTING.md` to reflect the changes in policies. I've added a single template to help gather the appropriate data and remind users of our requirements for code changes. I've created these after reviewing the pull request history for this repo as well as the pull request templates configured across other active projects.

This template will be listed as an option when a user opens a new pull request via the github ui, as outlined in the github documentation. We can create a directory with a different template for the different categories of pull requests, similar to the issue templates if we determine it's necessary.

* Dropped issue ref and squash requirement from PR

Addressing the comments during code review:
* No longer requiring the issue number in the title of the PR since we include a link to the related issues in the pull request template
* Dropping the comment about squashing related commits since we want to see the history and we'll do a squash-and-merge when it's approved

* Removed pull request checklist from CONTRIBUTING

This content shouldn't be maintained in two places and the focus here should really be on the green CI tests and the adding of tests to cover new functionality. The rest of this content is really more appropriate and relevant to be in the pull request template itself.

* Removing code conventions comment

This really should be in a linter and we shouldn't put the onus on our contributors for our own lack of tooling.

* Removing changelog specification requirement

I think we can do this via automation during the release process and we can agree on how we want to recognize contributors, highlight new rules, any significant bug fixes, or other changes that occurred to the project since the last release.