Better error messages for styles; better docs (#302)
This makes it clear to the user what went wrong and why:

```
$ ./bin/mdl -s 'test'
test does not appear to be a build-in style. If you meant to pass in your own style file, it must contain a '/' or end in '.rb'. See https://github.com/markdownlint/markdownlint/blob/master/docs/configuration.md
```

```
$ ./bin/mdl -s test.rb
Style 'test.rb' does not exist.
```

Closes #290

Signed-off-by: Phil Dibowitz <phil@ipom.com>