Enforce non-wrapping (#317)
I’ve added configuration to MD012 to switch it to enforce not wrapping to allow us to enforce a consistent approach. This is described in https://cirosantilli.com/markdown-style-guide/#option-wrap-no

I’m not a ruby dev but this seems to work, but please let me know where I’ve not followed convention or where there are tricks to make things read better.

I’ve created the test cases I can think of. I’m not happy with adding to overlines as these are not lines that are ‘over’. Perhaps this should be renamed something line violation_lines?

I also think there’s a fair bit of duplication in this function now, and some in the file in general around extracting line numbers for various things. I tried extracting a helper function in the test but this didn’t work. I was wondering about adding some more functionality to doc.rb but wanted to check with others first.