Fix `--git-recurse` (#340)
0.10.0 ran rubocopy autocorrect across the codebase.

In the process it replaced:

```ruby
%x(git ls-files '*.md' '*.markdown').split("\n")
```

with

```ruby
Mixlib::ShellOut.new("git ls-files '*.md' '*.markdown'").run_command.stdout.lines
```

Moving away from %x{} and to Mixlib::ShellOut is great! But `.lines` is *not* the same as `.split("\n")`.

`.lines` leaves newlines at the end of each entry in the array where as splitting on newlines does not. So strip the entries.

Fixes #339

Signed-off-by: Phil Dibowitz <phil@ipom.com>