docker: Create user with `adduser` command
Currently the necessary user within the image is created by modifying
the `/etc/passwd` and `/etc/group` files directly. While this is
sufficient to have a UID and GID, it misses some other aspects of user
creation, such as creating the home directory.

Instead of taking all of the necessary steps manually, it is recommended
to use the dedicated program for that, i.e. `adduser`. This way, all
necessary steps (including home directory creation) will be taken care of.

This addresses the issue reported in #361.

NOTE: Before the UID (and GID) was specifically set to 1000. Now, it
is no longer set explicitly and managed by the `adduser` command itself.
It happens to be 1000, since it is the first user. This matches the
previous behaviour, so from a user perspective nothing has changed.

Signed-off-by: Karol Babioch <karol@babioch.de>