apply markdownlint on RULES.md
The linter (version 0.12.0) prompts 100 suggestions if running on
the previous commit of RULES.md.  Though not applicable in every
instance showcasing the potential improvement (rules MD031, 040,
and 046 were not touched at all), it was possible to lower this
to 18, less than a fifth of the former.  Beside leading by example,
this causes the display of RULES.md to engage a syntax highlighting
similiar to the one one may encounter in an editor, too.