apply markdownlint on RULES.md (#439)
The linter (version 0.12.0) prompts 100 suggestions if running on
the previous commit of RULES.md.  Though not applicable in every
instance showcasing the potential improvement (rules MD031, MD040,
and MD046 were not touched at all), it was possible to lower this
to 18 instances, less than a fifth of the former; mainly by

* code fencing wit an explicit label `markdown`
* adjusting the indent of the unsorted list to three spaces
* where possible, to constrain lines to 80 characters per line (max)

Beside leading by example, this causes the display of RULES.md in
GitHub to engage a syntax highlighting similar to the one one may
encounter in an editor, too.

Signed-off-by: Norwid Behrnd <nbehrnd@yahoo.com>
Signed-off-by: iafelix <njr.iafelix@gmail.com>