Allow exactly 2 trailing spaces by default as per MD009 (#195) (#452)
The br_spaces parameter allows an exception to this rule for a specific
number of trailing spaces, typically used to insert an explicit line
break. The default value allows 2 spaces to indicate a hard break (<br>
element).

This default is more sensible in line with real world usage where two
spaces are rendered as hard break and not considered to be a format
issue.  Changing the default from 0 to 2 will not create undesired
surprises downstream as linting and CI of currently valid documents
won't be altered as the new default is more relaxed in that regard.

The new default of 2 br_spaces is also in line with other tools, linter
and renderer.

Signed-off-by: Levente Polyak <levente@leventepolyak.net>